+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project are documented below.
 The format is based on [keep a changelog](http://keepachangelog.com) and this project uses [semantic versioning](http://semver.org).
 
 ## [Unreleased]
+### Added
+- Added Nakama API to get users by facebook ID.
+
 ### Fixed
 - Fix an issue that prevented the JavaScript runtime hooks to be invoked correctly.
 - Fix the delete button not working in the console leaderboard listing.